requirements.txt
selenium~=4.27.1
pytest
requests
allure-pytest
pytest-html
webdriver-manager~=4.0.2